fabric-hr-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
msal
azure-identity
azure-graphrbac
pydantic
pydantic-settings
requests